+++ x/session/types/errors.go
@@ -8,8 +8,8 @@ import (
 
 // x/session module sentinel errors
 var (
-	ErrInvalidSigner                 = sdkerrors.Register(ModuleName, 1100, "expected gov account as only signer for proposal message")
-	ErrSample                        = sdkerrors.Register(ModuleName, 1101, "sample error")
+	ErrSessionInvalidSigner          = sdkerrors.Register(ModuleName, 1100, "expected gov account as only signer for proposal message")
+	ErrSessionSample                 = sdkerrors.Register(ModuleName, 1101, "sample error")
 	ErrSessionHydration              = sdkerrors.Register(ModuleName, 1102, "error during session hydration")
 	ErrSessionAppNotFound            = sdkerrors.Register(ModuleName, 1103, "application for session not found not found ")
 	ErrSessionAppNotStakedForService = sdkerrors.Register(ModuleName, 1104, "application in session not staked for requested service")
