--- x/session/types/genesis_test.go
@@ -1,41 +1,0 @@
-package types_test
-
-import (
-	"testing"
-
-	"github.com/pokt-network/poktroll/x/session/types"
-	"github.com/stretchr/testify/require"
-)
-
-func TestGenesisState_Validate(t *testing.T) {
-	tests := []struct {
-		desc     string
-		genState *types.GenesisState
-		valid    bool
-	}{
-		{
-			desc:     "default is valid",
-			genState: types.DefaultGenesis(),
-			valid:    true,
-		},
-		{
-			desc:     "valid genesis state",
-			genState: &types.GenesisState{
-
-				// this line is used by starport scaffolding # types/genesis/validField
-			},
-			valid: true,
-		},
-		// this line is used by starport scaffolding # types/genesis/testcase
-	}
-	for _, tc := range tests {
-		t.Run(tc.desc, func(t *testing.T) {
-			err := tc.genState.Validate()
-			if tc.valid {
-				require.NoError(t, err)
-			} else {
-				require.Error(t, err)
-			}
-		})
-	}
-}
